Added sprite support with 16x16 bitmap textures.
Somehow this actually worked on the first try, :O
To make a texture, just put it in the "tex" folder in the way that I did
the sword texture. Then add it into the init_textures method in
textures.c

Do not forget to track changes in the other files too now. If you add a
file, you need to do git add; the same goes for if you modify textures.h
or textures.c

To make a sprite, just declare it in the beginning of the main method
like I did with the sword sprite.